flask_crud_clientes/requirements.txt
flask
flask_sqlalchemy
psycopg2
flask_migrate
flask_wtf
python-dotenv
requests